+++ src/content/lib/request-processor.compat.js
@@ -89,6 +89,7 @@ RequestProcessor = (function(self) {
     idArray.push("{6614d11d-d21d-b211-ae23-815234e1ebb5}"); // Dr. Web Link Checker
     idArray.push("keefox@chris.tomlinson"); // KeeFox
     idArray.push("jid1-TPTs1Z1UvUn2fA@jetpack"); // Enpass
+    idArray.push("{0AA9101C-D3C1-4129-A9B7-D778C6A17F82}"); // rikaichan
 
     for (let id of idArray) {
       Logger.info(Logger.TYPE_INTERNAL, "Extension check: " + id);
@@ -200,6 +201,12 @@ RequestProcessor = (function(self) {
           "ws://localhost",
           ext.name
         ]);
+        break;
+
+      case "{0AA9101C-D3C1-4129-A9B7-D778C6A17F82}": // rikaichan
+        Logger.info(Logger.TYPE_INTERNAL,
+            "Using extension compatibility rules for: " + ext.name);
+        compatibilityRules.push([null, "chrome://rikaichan/skin/", ext.name]);
         break;
 
       default: